[ADD] Widget Numeric step
