--- web_widget_number_ux_choice/README.rst
@@ -1,86 +1,0 @@
-===========================
-Web Widget UX Number Choice
-===========================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-quentinDupont%2Fgrap--odoo--incubator-lightgray.png?logo=github
-    :target: https://github.com/quentinDupont/grap-odoo-incubator/tree/12.0_ADD_web_widget_int_choice/web_widget_number_ux_choice
-    :alt: quentinDupont/grap-odoo-incubator
-
-|badge1| |badge2| |badge3| 
-
-This widget changes input number field and make it easier to incremente the number thanks to 2 buttons (+ and -).
-We can be parametred to choose the iteration step (0.1 for more precision etc.)
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Usage
-=====
-
-In your xml view, add ``widget="number_ux_choice"``
-This will add the 2 buttons +" and "-" just next to the input field in edit mode.
-Iteration step by default is 1.
-
-.. figure:: https://raw.githubusercontent.com/quentinDupont/grap-odoo-incubator/12.0_ADD_web_widget_int_choice/web_widget_number_ux_choice/static/description/add_two_buttons.png
-
-
-**Optional**
-Add an option to choose the step iteration
-Example for an 0.1 step :
-
-``<field name="sale_delay" widget="number_ux_choice" options="{'step': 0.1}" /> days``
-
-**Examples**
-
-Iteration with 0.1 step
-
-.. figure:: https://raw.githubusercontent.com/quentinDupont/grap-odoo-incubator/12.0_ADD_web_widget_int_choice/web_widget_number_ux_choice/static/description/step0,1.gif
-
-Iteration with 10 step, with onchange
-
-.. figure:: https://raw.githubusercontent.com/quentinDupont/grap-odoo-incubator/12.0_ADD_web_widget_int_choice/web_widget_number_ux_choice/static/description/step10_with_onchange.gif
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/quentinDupont/grap-odoo-incubator/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/quentinDupont/grap-odoo-incubator/issues/new?body=module:%20web_widget_number_ux_choice%0Aversion:%2012.0_ADD_web_widget_int_choice%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* GRAP
-
-Contributors
-~~~~~~~~~~~~
-
-* `GRAP <http://www.grap.coop>`_:
-
-  * Quentin DUPONT <quentin.dupont@grap.coop>
-
-Maintainers
-~~~~~~~~~~~
-
-This module is part of the `quentinDupont/grap-odoo-incubator <https://github.com/quentinDupont/grap-odoo-incubator/tree/12.0_ADD_web_widget_int_choice/web_widget_number_ux_choice>`_ project on GitHub.
-
-You are welcome to contribute.
