--- web_widget_number_ux_choice/static/src/scss/number_ux_choice.scss
@@ -1,11 +1,0 @@
-
-.input_number_ux_choice {
-    width: 71% !important;
-}
-
-.btns_number_ux_choice {
-    width: 27% !important;
-    button {
-        height: 25px;
-    }
-}
