--- web_widget_number_ux_choice/view/web_widget_number_ux_choice.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<odoo>
-    <template id="assets_backend" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <link rel="stylesheet" href="/web_widget_number_ux_choice/static/src/scss/number_ux_choice.scss"/>
-            <script type="text/javascript" src="/web_widget_number_ux_choice/static/src/js/number_ux_choice.js"></script>
-        </xpath>
-    </template>
-</odoo>
