+++ web_widget_numeric_step/README.rst
@@ -1,0 +1,90 @@
+=======================
+Web Widget Numeric Step
+=======================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-quentinDupont%2Fweb-lightgray.png?logo=github
+    :target: https://github.com/quentinDupont/web/tree/12.0_ADD_web_widget_ux_choice/web_widget_numeric_step
+    :alt: quentinDupont/web
+
+|badge1| |badge2| |badge3| 
+
+This widget changes input number field and make it easier to incremente the number thanks to 2 buttons (+ and -).
+Use JS native logic for input number, so you can use the options ``min``, ``max``, ``step``, ``placeholder``.
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Usage
+=====
+
+In your xml view, add ``widget="numeric_step"``
+This will add the 2 buttons "+" and "-" just next to the input field in edit mode.
+Iteration step by default is 1.
+
+.. figure:: https://raw.githubusercontent.com/quentinDupont/web/12.0_ADD_web_widget_ux_choice/web_widget_numeric_step/static/description/add_two_buttons.png
+
+
+**Optional**
+
+Add an option to choose the step iteration and limits (min and max values).
+
+Example for an 0.25 step, min to -1 and max to 10 :
+
+``<field name="sale_delay" widget="numeric_step" options="{'step': 0.25, 'min': -1, 'max': 10}" /> days``
+
+**Examples**
+
+Iteration with 0.25 step, min to -1 and max to 10.
+
+Start to incremente with button, continue incrementing with scrolling mouse.
+
+.. figure:: https://raw.githubusercontent.com/quentinDupont/web/12.0_ADD_web_widget_ux_choice/web_widget_numeric_step/static/description/step0,25andlimits.gif
+
+Iteration with 10 step, max limit 15, placeholder with onchange
+
+.. figure:: https://raw.githubusercontent.com/quentinDupont/web/12.0_ADD_web_widget_ux_choice/web_widget_numeric_step/static/description/step10_limit15_placeholder117_with_onchange.gif
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/quentinDupont/web/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/quentinDupont/web/issues/new?body=module:%20web_widget_numeric_step%0Aversion:%2012.0_ADD_web_widget_ux_choice%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* GRAP
+
+Contributors
+~~~~~~~~~~~~
+
+* `GRAP <http://www.grap.coop>`_:
+
+  * Quentin DUPONT <quentin.dupont@grap.coop>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `quentinDupont/web <https://github.com/quentinDupont/web/tree/12.0_ADD_web_widget_ux_choice/web_widget_numeric_step>`_ project on GitHub.
+
+You are welcome to contribute.
