+++ web_widget_numeric_step/__manifest__.py
@@ -1,8 +1,8 @@
-# Odoo, Open Source Web Widget Integer Choice
+# Odoo, Open Source Web Widget Numeric Step
 #
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).#
 {
-    'name': "Web Widget UX Number Choice",
+    'name': "Web Widget Numeric Step",
     'category': "web",
     'version': "12.0.1.0.0",
     'author': "GRAP, "
@@ -11,10 +11,10 @@
     'website': 'https://github.com/OCA/web',
     'depends': ['web'],
     'data': [
-        'view/web_widget_number_ux_choice.xml'
+        'view/web_widget_numeric_step.xml'
     ],
     'qweb': [
-        'static/src/xml/number_ux_choice.xml',
+        'static/src/xml/numeric_step.xml',
     ],
     'auto_install': False,
     'installable': True,
