+++ web_widget_numeric_step/i18n/fr.po
