+++ web_widget_numeric_step/readme/DESCRIPTION.rst
@@ -1,2 +1,2 @@
 This widget changes input number field and make it easier to incremente the number thanks to 2 buttons (+ and -).
-We can be parametred to choose the iteration step (0.1 for more precision etc.)
+Use JS native logic for input number, so you can use the options ``min``, ``max``, ``step``, ``placeholder``.
