+++ web_widget_numeric_step/static/description/add_two_buttons.png
