+++ web_widget_numeric_step/static/description/icon.png
