+++ web_widget_numeric_step/static/src/scss/numeric_step.scss
@@ -1,0 +1,29 @@
+
+.input_numeric_step {
+    width: 71% !important;
+}
+
+.btns_numeric_step {
+    width: 27% !important;
+    button {
+        height: 25px;
+    }
+}
+
+/* Turn Off Number Input Spinners */
+input{
+    /* Firefox */
+    -moz-appearance: textfield;
+
+    /* Chrome */
+    &::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin:0;
+    }
+
+    /* Opéra*/
+    &::-o-inner-spin-button {
+        -o-appearance: none;
+        margin: 0; 
+    }
+}
