+++ web_widget_numeric_step/static/src/xml/numeric_step.xml
@@ -5,11 +5,11 @@ Copyright (C) 2019 - Today: GRAP (http://www.grap.coop)
 License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 -->
 <templates>
-    <t t-name="NumberUxChoice">
+    <t t-name="web_widget_numeric_step">
         <t t-if="widget.mode != 'readonly'">
-            <input class="input_number_ux_choice"/>
-            <div class="btns_number_ux_choice">
-                <button class="fa fa-plus btn btn-secondary btn_number_ux_choice_plus" aria-label="Plus" title="Plus"/><button class="fa fa-minus btn btn-secondary btn_number_ux_choice_minus" aria-label="Minus" title="Minus"/>
+            <input type="number" class="input_numeric_step"/>
+            <div class="btns_numeric_step">
+                <button class="fa fa-plus btn btn-secondary btn_numeric_step_plus" aria-label="Plus" title="Plus"/><button class="fa fa-minus btn btn-secondary btn_numeric_step_minus" aria-label="Minus" title="Minus"/>
             </div>
         </t>
         <t t-if="widget.mode == 'readonly'">
